Fix circular class loading
